license_plate_recognition/ocrcar/requirements.txt
einops==0.4.0
Flask==2.0.2
matplotlib==3.3.0
numpy==1.20
opencv-python==4.2.0.34
Pillow==8.4.0
pyparsing==3.0.7
PyYAML==5.3.1
qrcode==7.3.1
requests==2.24.0
seaborn==0.11.2
six==1.16.0
tensorboard==2.11
thop==0.0.31.post2005241907
urllib3==1.25.11
Werkzeug==2.0.1
